packages/note-chat-server/requirements.txt
langchain==0.0.58
openai
faiss-cpu
streamlit
streamlit-chat
Flask==2.2.2
tiktoken
flask_cors
gunicorn
